requirements.txt
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0
PyQt6-sip>=13.6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
dnspython>=2.4.0
cryptography>=41.0.0
pyjwt>=2.8.0
PyYAML>=6.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
httpx>=0.25.0
python-nmap>=0.7.1
paramiko>=3.3.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0
pymongo>=4.5.0
redis>=5.0.0
celery>=5.3.0
flower>=2.0.0
jinja2>=3.1.0
markdown>=3.5.0
reportlab>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
Pillow>=10.0.0
qrcode>=7.4.0
pyperclip>=1.8.0
colorama>=0.4.6
tqdm>=4.66.0
python-dateutil>=2.8.0
pytz>=2023.3
validators>=0.22.0
urllib3>=2.0.0
websocket-client>=1.6.0
mitmproxy>=10.1.0
capstone>=5.0.0
keystone-engine>=0.9.2
pefile>=2023.2.7
